+++ app/service/roles.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const Service = require('egg').Service;
+const { Op } = require('sequelize');
 
 /**
  * @service 角色 roles
@@ -42,6 +43,18 @@ class RolesService extends Service {
         if (!allRoles.every(e => e.is_default !== 1)) {
             return { __code_wrong: 40000 };
         }
+        // 角色已经被用户使用，不允许删除
+        const relatedData = await ctx.model.UserRoles.findAndCountAll(
+            {
+                where: { role_id: payload.ids },
+            }
+        );
+        // 在删除资源时，如果有角色关联了该资源，则不允许删除
+        if (relatedData.count > 0) {
+            return {
+                __code_wrong: 40001
+            }
+        }
         return await ctx.model.Roles.destroy({ where: { id: payload.ids } });
     }
 
@@ -62,12 +75,12 @@ class RolesService extends Service {
         await ctx.model.Roles.update({ is_default: 0 }, { where: { is_default: 1 }, transaction });
         const res = await ctx.model.Roles.update({ is_default: 1 }, { where: { id: payload.id }, transaction });
         if (res && res[0] === 1) {
-          await transaction.commit();
-          return true;
+            await transaction.commit();
+            return true;
         }
         await transaction.rollback();
         return false;
-      }
+    }
 }
 
 module.exports = RolesService;
